Move getSymbolParams method to helper
