+++ workers/loc.api/helpers/prepare-response/helpers/index.js
@@ -5,11 +5,13 @@ const getParamsSchemaName = require('./get-params-schema-name')
 const omitPrivateModelFields = require('./omit-private-model-fields')
 const getBfxApiMethodName = require('./get-bfx-api-method-name')
 const getSymbolsForFiltering = require('./get-symbols-for-filtering')
+const getSymbolParams = require('./get-symbol-params')
 
 module.exports = {
   getParamsMap,
   getParamsSchemaName,
   omitPrivateModelFields,
   getBfxApiMethodName,
-  getSymbolsForFiltering
+  getSymbolsForFiltering,
+  getSymbolParams
 }
